feat(#26): create NutrientDailyValue type and extract generateDailyValues in shared space
