+++ src/templates/usda/FoundationFoodNutritionFacts.tsx
@@ -1,31 +1,13 @@
 import * as React from "react"
 import { Breadcrumbs } from "../../components/Breadcrumb"
-import { FoundationFood, RDI } from "../../shared/types"
-import {
-  getDailyPercentValue,
-  mappingsByNutrient,
-} from "../../shared/functions"
+import { FoundationFood, NutrientDailyValue, RDI } from "../../shared/types"
+import { generateDailyValues } from "../../shared/functions"
 
 export default ({ pageContext }) => {
   const { food, rdis, breadcrumbs } = pageContext
   const thisFood = food as FoundationFood
-  const thisRdis = rdis as RDI[]
-  const generateDailyValues = () => {
-    return thisFood.nutrients.map(nutrient => {
-      const mappedRdi = mappingsByNutrient.get(nutrient.name)
-
-      if (!mappedRdi) return { nutrient }
-
-      const rdisForNutrient = thisRdis.filter(
-        rdi => rdi.nutrient === mappedRdi.rdiNutrientName
-      )
-      return rdisForNutrient.map(rdi => {
-        const percentDaily = getDailyPercentValue(nutrient, rdi)
-        return { nutrient, rdi, percentDaily }
-      })
-    })
-  }
-  const values = generateDailyValues()
+  const allRdis = rdis as RDI[]
+  const values: NutrientDailyValue[] = generateDailyValues(thisFood, allRdis)
 
   return (
     <div
@@ -47,7 +29,7 @@ export default ({ pageContext }) => {
       <div>
         <h1>RDI</h1>
         <pre style={{ height: 600, width: "auto" }}>
-          {JSON.stringify(thisRdis, null, 2)}
+          {JSON.stringify(allRdis, null, 2)}
         </pre>
       </div>
       <div>
